Factor out plugins.
